+++ src/message.c
@@ -1708,7 +1708,11 @@ static void handle_domain_display(FILE *rsp, struct token domain, char *message)
     if (token_equals(command, COMMAND_DISPLAY_FOCUS)) {
         struct selector selector = parse_display_selector(rsp, &message, acting_did, false);
         if (selector.did_parse && selector.did) {
-            display_manager_focus_display(selector.did, display_space_id(selector.did));
+            if (acting_did != selector.did) {
+                display_manager_focus_display(selector.did, display_space_id(selector.did));
+            } else {
+                daemon_fail(rsp, "cannot focus an already focused display.\n");
+            }
         }
     } else if (token_equals(command, COMMAND_DISPLAY_SPACE)) {
         struct selector selector = parse_space_selector(rsp, &message, display_space_id(acting_did), false);